Rename the debian-etch-debootstrap OS definition
The current arrangement of the subversion tree is not good - the OS
definitions are treated differently from the ganeti core code, and
automatic operations against the tree are not as easy to do as if the
model would be the same.

This patch moves the trunk/os-scripts/debian-etch-debootstrap directory
to trunk/instance-debootstrap; the prefix 'instance' denotes this as an
OS definition, and we intend to make this OS more flexible so we drop
the debian and etch parts.

Reviewed-by: ultrotter, imsnah